refactor: RemoteError - HttpException 상속으로 변경
